tests/wpt/tests/webmessaging/with-ports/023.html

<!doctype html>
<title>null ports</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
test(function(t) {
  assert_throws_js(TypeError, () => postMessage('', '*', null));
});
</script>
